Move `watch` command from CLI dashboard module to deprecated module
Simplify imports in the CLI entry point by importing `watch` from
`swanlab.deprecated.local` instead of the deleted dashboard submodule.
