+++ swanlab/deprecated/local.py
@@ -11,6 +11,8 @@ import math
 import os
 import re
 import shutil
+import socket
+import sys
 import tempfile
 import time
 from datetime import datetime, timezone
@@ -19,8 +21,155 @@ from importlib.metadata import version
 from pathlib import Path
 from typing import Any, Optional, TextIO, cast
 
+import click
+
+from swanlab.sdk import settings
+from swanlab.sdk.internal import pkg
 from swanlab.sdk.internal.settings import Settings
 from swanlab.sdk.protocol.callbacker import Callback
+
+
+def _get_free_port(address: str = "0.0.0.0", default_port: int = 5092) -> int:
+    """获取一个可用端口。默认返回 5092，如果被占用，返回一个随机可用端口。"""
+    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
+        try:
+            s.bind((address, default_port))
+        except OSError:
+            pass
+        else:
+            return default_port
+    print(f"Port {default_port} is occupied, trying to find a free port...")
+    sock = socket.socket()
+    sock.bind((address, 0))
+    _, port = sock.getsockname()
+    sock.close()
+    return port
+
+
+@click.command()
+@click.argument(
+    "path",
+    type=click.Path(
+        exists=True,
+        dir_okay=True,
+        file_okay=False,
+        resolve_path=True,
+        readable=True,
+    ),
+    nargs=1,
+    required=False,
+    default=None,
+)
+@click.option(
+    "--host",
+    "-h",
+    default=settings.integration.dashboard.host,
+    type=str,
+    nargs=1,
+    help="The host of swanlab web, default by 127.0.0.1",
+)
+@click.option(
+    "--port",
+    "-p",
+    default=settings.integration.dashboard.port,
+    nargs=1,
+    type=click.IntRange(1, 65535),
+    help="The port of swanlab web, default by 5092",
+)
+@click.option(
+    "--log-dir",
+    "-l",
+    default=None,
+    nargs=1,
+    type=click.Path(
+        exists=True,
+        dir_okay=True,
+        file_okay=False,
+        resolve_path=True,
+        readable=True,
+    ),
+    help="Specify the folder to store Swanlog. Deprecated: use `swanlab watch <LOG PATH>` instead.",
+)
+@click.option(
+    "--logdir",
+    default=None,
+    nargs=1,
+    type=click.Path(
+        exists=True,
+        dir_okay=True,
+        file_okay=False,
+        resolve_path=True,
+        readable=True,
+    ),
+    help="Deprecated: use --log-dir instead.",
+    hidden=True,
+)
+def watch(path: str, host: str, port: int, log_dir: str, logdir: str):
+    """Run this command to turn on the SwanLab dashboard service."""
+    # logdir 兼容旧参数（向后兼容）
+    if logdir is not None:
+        click.echo("Warning: The option `--logdir` is deprecated, use `--log-dir` instead.")
+        log_dir = logdir
+
+    # log-dir 覆盖 path（向后兼容）
+    if log_dir is not None:
+        click.echo(
+            "Warning: The option `--log-dir` is deprecated, use `swanlab watch [PATH]` to specify the path instead."
+        )
+        path = log_dir
+
+    if path is not None:
+        path = os.path.abspath(path)
+
+    port = _get_free_port(default_port=port)
+
+    try:
+        # noinspection PyPackageRequirements
+        from swanboard import SwanBoardRun
+        from swanboard.utils import get_swanlog_dir
+    except ModuleNotFoundError:
+        raise
+        click.echo("Please install the swanboard package: `pip install swanlab[dashboard]`")
+        return sys.exit(1)
+    # ----- 校验path，path如果被输入，已经由上层校验已存在，可读，是一个文件夹 -----
+    if logdir is not None:
+        pkg.console.warning(
+            "The option `--logdir` will be deprecated in the future, "
+            "you can just use `swanlab watch [PATH]` to specify the path."
+        )
+    # logdir 覆盖 path，接下来统一处理path而不再管logdir
+    path = logdir if logdir is not None else path
+    if path is not None:
+        path = os.path.abspath(path)
+        os.environ["SWANLAB_LOG_DIR"] = path
+    # 为None时从环境变量中获取
+    try:
+        path = get_swanlog_dir()
+        # 产品经理要求无论日志文件夹是否存在都不报错 🤡 ，给用户以开启web服务的“爽感”
+        # if not os.path.exists(path):
+        #     raise FileNotFoundError
+    except ValueError as e:
+        click.BadParameter(str(e))
+        return sys.exit(3)
+    except NotADirectoryError:
+        click.BadParameter("SWANLAB_LOG_DIR must be a directory")
+        return sys.exit(4)
+    except FileNotFoundError:
+        click.BadParameter(f"The log folder `{path}` was not found")
+        return sys.exit(5)
+    # ----- 校验host和port -----
+    try:
+        SwanBoardRun.is_valid_port(port)
+        SwanBoardRun.is_valid_ip(host)
+    except ValueError as e:
+        click.BadParameter(str(e))
+        return sys.exit(6)
+    # ---- 启动服务 ----
+    SwanBoardRun.run(
+        path=path,
+        host=host,
+        port=port,
+    )
 
 
 class _ChartTypeItem:
@@ -197,10 +346,11 @@ class LocalCallbacker(Callback):
         except ImportError as e:
             raise ImportError("Please install swanboard to use 'local' mode: pip install 'swanlab[dashboard]'") from e
 
-        if version("swanboard") != "0.1.10b0":
+        if os.getenv("SWANLAB_SKIP_SWANBOARD_VERSION_CHECK") != "1" and version("swanboard") != "0.1.10b2":
             raise ImportError(
                 "Your swanboard version does not match, please use this command to install the matching version: "
-                "pip install 'swanlab[dashboard]'"
+                "pip install 'swanlab[dashboard]'. "
+                "To skip this check temporarily, set SWANLAB_SKIP_SWANBOARD_VERSION_CHECK=1."
             )
 
         get_swanlog_dir = getattr(import_module("swanboard.utils"), "get_swanlog_dir")
@@ -415,12 +565,22 @@ class LocalCallbacker(Callback):
         legacy_dir = media_dir / ctx.kid
         legacy_dir.mkdir(parents=True, exist_ok=True)
         for item in record.value.items:
-            filenames.append(item.filename)
-            more.append(None)
             source = media_dir / medium / item.filename
-            target = legacy_dir / item.filename
-            self._ensure_legacy_media_link(source, target)
+            if medium == "text":
+                filenames.append(self._read_text_media(source, item.filename))
+            else:
+                filenames.append(item.filename)
+                target = legacy_dir / item.filename
+                self._ensure_legacy_media_link(source, target)
+            more.append({"caption": item.caption} if item.caption else None)
         return filenames, more
+
+    @staticmethod
+    def _read_text_media(source: Path, fallback: str) -> str:
+        try:
+            return source.read_text(encoding="utf-8")
+        except OSError:
+            return fallback
 
     @staticmethod
     def _ensure_legacy_media_link(source: Path, target: Path) -> None:
